+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 
 This document records the main changes to the ``sdssdb`` code.
 
+* Add table and models for ``lvmopsdb.ln2_fill``.
+
 * :release:`0.13.0 <2024-09-03>`
 * :feature:`264` Add metadata for the ``sdss_id_to_catalog`` table.
 * :feature:`266` Support PEP 621.